See below
Updated the home page for the Oct 2019 comp.  Also updated the results page so that visitors know that I'm awaiting the results.  In addition, I've added a field which allows me to enter the date that I contacted Julie for the results, so that visitors don't think that nothing has been happening.